Akaike Information Criterion With Correction

The function calculates the Akaike Information Criterion with correction for small samples size.

Usage

AICC(fit)

Arguments

fit

a 'nls'-object

Value

Returns the AICc value

Details

When the sample size is small, there is a substantial probability that AIC (see stats::AIC() for more details) will select models that have too many parameters, i.e. that AIC will overfit. AICc is AIC with a correction for small sample sizes.

The AICc is computed as follows: $$AICc=AIC+\frac{2\,k\,(k+1)}{n-k-1}$$ where n denotes the sample size and k denotes the number of parameters. Thus , AICc is essentially AIC with an extra penalty term for the number of parameters. Note that as \(n\rightarrow \infty\), the extra penalty term converges to 0, and thus AICc converges to AIC.

See also

Examples

t <- seq(0, 10, 1)
y <- 1 / (0.5 * exp(t) + 1) + stats::rnorm(length(t), 0, 0.05)

fit <- nls(y ~ 1 / (k * exp(t) + 1),
  data = list(t = t, y = y),
  start = list(k = 0.2)
)
AICC(fit)
#> [1] -34.94881